Resolve Month Page
We used to use something called a StringDictionary and ArrayList to
track the post for the year. I assume the Dictionary/ArrayList object
kept the year unique and thus limited the page to one year regardless of
the posts for that year. A more recent update changed the
StringDictionary to a HashSet and ArrayList to Arrary, then never used
the hashset.

Closes #8