Use the content to retrieve blog post details
Folks usually use the "Description" property to store a summary of the
actual post. You could miss the images and videos you want to include in
the meta tags by using description. Updated to use the posts "Content"
for the basis of the search for images/video.

Also renamed a couple of variables.